Issue #5 -- Give UserRelationships their own state and resource.
This is to fix the bug in friends.  Keeping UserRelationships as a
special snowflake that lives on the user session is the root cause of
several bugs.  Giving UserRelationships its own state and REST resource
will help resolve those bugs and keep more consistent patterns. There's
really no reason not to.  I tied myself in a bit of a knot when I
originally decided not to.