Issue #19 -- Refactoring PermissionService and using it in GroupMemberController.
Refactoring PermissionService to use any context it is given and only
query for context if it doesn't already have it.  That way we can query
for the context ahead of time when we know we need to do a bunch of
related permissions checks.

Refactoring GroupMemberController to use PermissionService for
permissions checks.