fix: patch user on submit so future file uploads don't auto-patch
When the user submitted a new profile picture we were previously
patching the user only after the profile picture was cropped using a
useEffect hook.  The problem is we weren't cleaning up the file request,
so that hook remained active if the user uploaded a second image and
would cause the user to be immediately patched with the new image,
instead of waiting for the form to submit.

We don't need to wait for the file to crop to attach it to the user.
Now when you crop a file, the file and patch request are submitted
simultaneously.