refactor: permissions are not longer a hook, context pulled separately
Removing `useGroupPermission()`, `useGroupMemberPermission()`, and
`useSitePermission()` in favor of the non-hook `can()`.  The reasoning
for making them a hook was that the permission hooks could retrieve
their own context, but that doesn't actually work with react.  It
results in tons of duplicate queries.

So instead, the context gathering was pulled into
`useGroupPermissionContext()` and the permission calculations was put
into `can()`.

Refs: #165