feat: Don't Keep Birthdates
We don't actually want to hang on to Birthdates right now.  It's
sensitive information that we never wanted to collect, we just had to
for the age gate.  But once we've validated someone's age, we don't need
the birthdate anymore.  So delete it from the database as soon as they
are validated.

Refs: #331