fix: cover all cases in FetchImage needToLoad
We were missing the cases in needToLoad where we don't have a width and
fall back to 'full'.  These were previously being covered by the
imageUrl being null, but since we're now using that to indicate that the
image is already being loaded we can't rely on that and need to cover
those cases in `needToLoad`.